Prefix environment variables with "DOCKERFILES_"
This is done due to how base image ENV and child image ARG collisions
work. When they collide, the base image ENV takes precidence instead
of allowing the child image ARG to override it. This sucks for images
that derive from these, so we deal with the pain of prefixing all of
our environment variables to make their lives easier.